+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All changes to this project will be documented in this file.
 
+## [1.0.9] - 2022-05-24
+- Progressive upload: prevent last part to be empty
+
 ## [1.0.8] - 2022-04-27
 - Create a AbstractUploader
 - Add origin header